@@ -19,40 +19,43 @@
     "babel-preset-react": "^6.0.0",
     "babel-runtime": "^6.26.0",
     "bootstrap": "^3.3.7",
-    "css-loader": "^0.9.1",
-    "file-loader": "^0.8.1",
-    "jsx-loader": "^0.12.2",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
-    "react-bootstrap": "^0.32.4",
-    "react-hot-loader": "^1.2.4",
-    "style-loader": "^0.8.3",
-    "url-loader": "^0.5.5",
+    "css-loader": "^3.5.0",
+    "html-webpack-plugin": "^3.0.0",
+    "js-cookie": "^2.2.1",
+    "jsx-loader": "^0.13.0",
+    "lodash": "^4.17.0",
+    "moment": "^2.22.2",
+    "node-sass": "^4.9.0",
+    "react": "^16.13.0",
+    "react-bootstrap": "^1.0.0",
+    "react-dom": "^16.12.0",
+    "react-hot-loader": "^4.12.18",
+    "react-notification-system": "^0.2.0",
+    "resolve-url-loader": "^2.3.0",
+    "sass-loader": "^8.0.0",
+    "scss": "^0.2.4",
+    "style-loader": "^1.2.0",
+    "url-loader": "^4.1.0",
     "webpack": "^4.0.0",
     "webpack-cli": "^3.0.0",
-    "webpack-dev-server": "^3.1.0",
-    "react-tag-input": "^4.7.2",
-    "html-webpack-plugin": "^3.0.0",
-    "react-notification-system": "^0.2.0",
-    "react-widgets": "^4.3.2",
-    "lodash": "^4.17.10",
-    "moment": "^2.22.2",
-    "react-widgets-moment": "^4.0.15",
-    "raw-loader": "^0.5.1",
-    "sass-loader": "^7.0.3",
-    "scss": "^0.2.4",
-    "node-sass": "^4.9.0",
-    "resolve-url-loader": "^2.3.0",
-    "js-cookie": "^2.2.1"
+    "webpack-dev-server": "^3.1.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "minimize": "webpack -p --progress --colors --devtool none",
     "zip": "zip -jr release.zip build",
-    "start": "node server.js"
+    "start": "node server.js --env.project=$REAPER_HOST",
+    "navbar": "./build-navbar.sh"
   },
   "dependencies": {
+    "core-js": "^2.6.11",
+    "create-react-class": "^15.6.3",
     "grunt": "^1.0.4",
-    "markdown-it": "^10.0.0"
+    "markdown-it": "^10.0.0",
+    "prop-types": "^15.7.2",
+    "react-datepicker": "^2.14.1",
+    "react-dnd": "^10.0.2",
+    "react-dnd-html5-backend": "^10.0.2",
+    "react-select": "^3.0.8"
   }
 }
