@@ -15,15 +15,17 @@
 //  limitations under the License.
 
 import React from "react";
+import CreateReactClass from'create-react-class';
+import PropTypes from 'prop-types';
 import moment from "moment";
 import {RowDeleteMixin, DeleteStatusMessageMixin} from "jsx/mixin";
 
-const TableRow = React.createClass({
+const TableRow = CreateReactClass({
   mixins: [RowDeleteMixin],
 
   propTypes: {
-    deleteSubject: React.PropTypes.object.isRequired,
-    listenSubscriptionSubject: React.PropTypes.object.isRequired
+    deleteSubject: PropTypes.object.isRequired,
+    listenSubscriptionSubject: PropTypes.object.isRequired
   },
 
   _onView: function() {
@@ -83,21 +85,21 @@ const TableRow = React.createClass({
   }
 });
 
-const eventSubscriptionList = React.createClass({
+const eventSubscriptionList = CreateReactClass({
   mixins: [DeleteStatusMessageMixin],
 
   propTypes: {
-    eventSubscriptions: React.PropTypes.object.isRequired,
-    deleteSubscriptionSubject: React.PropTypes.object.isRequired,
-    deleteResult: React.PropTypes.object.isRequired,
-    listenSubscriptionSubject: React.PropTypes.object.isRequired
+    eventSubscriptions: PropTypes.object.isRequired,
+    deleteSubscriptionSubject: PropTypes.object.isRequired,
+    deleteResult: PropTypes.object.isRequired,
+    listenSubscriptionSubject: PropTypes.object.isRequired
   },
 
   getInitialState: function() {
     return {eventSubscriptions: [], deleteResultMsg: null};
   },
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._subscriptionsSubscription = this.props.eventSubscriptions.subscribeOnNext(obs =>
       obs.subscribeOnNext(subscriptions => {
         this.setState({eventSubscriptions: subscriptions});
