@@ -14,28 +14,30 @@
 //  limitations under the License.
 
 import React from "react";
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
 import moment from "moment";
 import ServerStatus from "jsx/server-status";
 import Sidebar from "jsx/sidebar";
-import ScheduleForm from "jsx/schedule-form";
+import RepairForm from "jsx/repair-form";
 import ScheduleList from "jsx/schedule-list";
 import NavBar from "jsx/navbar";
 import {RowDeleteMixin, RowAbortMixin, StatusUpdateMixin, DeleteStatusMessageMixin, CFsListRender} from "jsx/mixin";
 
-const ScheduleScreen = React.createClass({
+const ScheduleScreen = CreateReactClass({
   mixins: [RowDeleteMixin, StatusUpdateMixin, RowAbortMixin],
 
   propTypes: {
-    schedules: React.PropTypes.object.isRequired,
-    clusterNames: React.PropTypes.object.isRequired,
-    deleteSubject: React.PropTypes.object.isRequired,
-    updateStatusSubject: React.PropTypes.object.isRequired,
-    deleteResult: React.PropTypes.object.isRequired,
-    currentCluster: React.PropTypes.string.isRequired,
-    statusObservableTimer: React.PropTypes.object.isRequired,
-    logoutSubject: React.PropTypes.object.isRequired,
-    logoutResult: React.PropTypes.object.isRequired,
-    switchTheme: React.PropTypes.func
+    schedules: PropTypes.object.isRequired,
+    clusterNames: PropTypes.object.isRequired,
+    deleteSubject: PropTypes.object.isRequired,
+    updateStatusSubject: PropTypes.object.isRequired,
+    deleteResult: PropTypes.object.isRequired,
+    currentCluster: PropTypes.string.isRequired,
+    statusObservableTimer: PropTypes.object.isRequired,
+    logoutSubject: PropTypes.object.isRequired,
+    logoutResult: PropTypes.object.isRequired,
+    switchTheme: PropTypes.func
   },
 
   getInitialState: function() {
@@ -74,7 +76,7 @@ const ScheduleScreen = React.createClass({
 
       <div className="row">
                 <div className="col-lg-12">
-                  <ScheduleForm addScheduleSubject={this.props.addScheduleSubject} addScheduleResult={this.props.addScheduleResult} clusterNames={this.props.clusterNames} currentCluster={this.props.currentCluster}> </ScheduleForm>
+                  <RepairForm addRepairSubject={this.props.addRepairSubject} addRepairResult={this.props.addRepairResult} clusterNames={this.props.clusterNames} currentCluster={this.props.currentCluster} formType="schedule"> </RepairForm>
                 </div>
             </div>
 
