@@ -14,35 +14,43 @@
 //  limitations under the License.
 
 import React from "react";
-import { WithContext as ReactTags } from 'react-tag-input';
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
+import Select from 'react-select';
 import {getUrlPrefix, toast} from "jsx/mixin";
 import $ from "jquery";
 var NotificationSystem = require('react-notification-system');
 
-const snapshotForm = React.createClass({
+const snapshotForm = CreateReactClass({
   _notificationSystem: null,
 
   propTypes: {
-    clusterNames: React.PropTypes.object.isRequired,
-    listSnapshots: React.PropTypes.func,
-    changeCurrentCluster: React.PropTypes.func
+    clusterNames: PropTypes.object.isRequired,
+    listSnapshots: PropTypes.func,
+    changeCurrentCluster: PropTypes.func
   },
 
   getInitialState: function() {
     const URL_PREFIX = getUrlPrefix(window.top.location.pathname);
     
     return {
-      clusterNames: [], submitEnabled: false,
-      clusterName: this.props.currentCluster!="all"?this.props.currentCluster:this.props.clusterNames[0], 
-      keyspace: "", tables: "", owner: "", name: "",
-      cause: "", formCollapsed: true, 
-      clusterStatus: {}, urlPrefix: URL_PREFIX, tableSuggestions: [], 
-      clusterTables: {}, tableList: [], keyspaceList: [], keyspaceSuggestions: [],
-      tablelistReadOnly: false
+      submitEnabled: false,
+      clusterNames: [],
+      clusterName: this.props.currentCluster === "all" ? this.props.clusterNames[0] : this.props.currentCluster,
+      keyspace: "",
+      owner: "",
+      name: "",
+      cause: "",
+      formCollapsed: true,
+      clusterStatus: {},
+      urlPrefix: URL_PREFIX,
+      clusterTables: {},
+      keyspaceList: [],
+      keyspaceOptions: [],
     };
   },
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._clusterNamesSubscription = this.props.clusterNames.subscribeOnNext(obs =>
       obs.subscribeOnNext(names => {
         let previousNames = this.state.clusterNames;
@@ -66,56 +74,60 @@ const snapshotForm = React.createClass({
   _getClusterStatus: function() {
     let clusterName = this.state.clusterName;
     $.ajax({
-          url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName),
-          method: 'GET',
-          component: this,
-          complete: function(data) {
-            this.component.setState({clusterStatus: $.parseJSON(data.responseText)});
-          }
-      });
+      url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName),
+      method: 'GET',
+      component: this,
+      complete: function(data) {
+        this.component.setState({clusterStatus: $.parseJSON(data.responseText)});
+      }
+    });
     $.ajax({
       url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName) + '/tables',
       method: 'GET',
       component: this,
       complete: function(data) {
         this.component.setState({clusterTables: $.parseJSON(data.responseText)});
-        this.component._getKeyspaceSuggestions();
+        this.component._getKeyspaceOptions();
       }
     });
   },
 
   _takeSnapshot: function(snapshot) {
-      this.setState({communicating: true}); 
-      toast(this._notificationSystem, "Taking a new snapshot...", "warning", snapshot.clusterName);  
-      const snapshotParams = '&snapshot_name=' + encodeURIComponent(snapshot.name==''?'reaper':snapshot.name) 
-                           + '&owner=' + encodeURIComponent(snapshot.owner)
-                           + '&cause=' + encodeURIComponent(snapshot.cause);
+    this.setState({communicating: true});
+    toast(this._notificationSystem, "Taking a new snapshot...", "warning", snapshot.clusterName);
+    const snapshotParams = '&snapshot_name='
+      + encodeURIComponent(snapshot.name === '' ? 'reaper' : snapshot.name)
+      + '&owner=' + encodeURIComponent(snapshot.owner)
+      + '&cause=' + encodeURIComponent(snapshot.cause);
       
-      $.ajax({
-            url: getUrlPrefix(window.top.location.pathname) + '/snapshot/cluster/' + encodeURIComponent(snapshot.clusterName) + "?keyspace=" + snapshot.keyspace + snapshotParams,
-            dataType: 'text',
-            method: 'POST',
-            component: this,
-            success: function(data) {
-                toast(this.component._notificationSystem, "Successfully took a new snapshot", "success", snapshot.clusterName); 
-            },
-            complete: function(data) {
-              this.component.setState({communicating: false});
-              this.component.props.changeCurrentCluster(snapshot.clusterName);
-              this.component.props.listSnapshots(snapshot.clusterName);
-            },
-            error: function(data) {
-                toast(this.component._notificationSystem, "Failed taking a snapshot on cluster " + snapshot.clusterName + " : " + data.responseText, "error", snapshot.clusterName);
-            }
-        });
+    $.ajax({
+      url: getUrlPrefix(window.top.location.pathname) + '/snapshot/cluster/' + encodeURIComponent(snapshot.clusterName) + "?keyspace=" + snapshot.keyspace + snapshotParams,
+      dataType: 'text',
+      method: 'POST',
+      component: this,
+      success: function(data) {
+        toast(this.component._notificationSystem, "Successfully took a new snapshot", "success", snapshot.clusterName);
+      },
+      complete: function(data) {
+        this.component.setState({communicating: false});
+        this.component.props.changeCurrentCluster(snapshot.clusterName);
+        this.component.props.listSnapshots(snapshot.clusterName);
+      },
+      error: function(data) {
+        toast(
+          this.component._notificationSystem,
+          "Failed taking a snapshot on cluster " + snapshot.clusterName + " : " + data.responseText,
+          "error",
+          snapshot.clusterName
+        );
+      }
+    });
   },
 
-  _getKeyspaceSuggestions: function() {
-    this.setState({keyspaceSuggestions: Object.keys(this.state.clusterTables)});
-  },
-
-  _getTableSuggestions: function(ks) {
-    this.setState({tableSuggestions: this.state.clusterTables[ks]});
+  _getKeyspaceOptions: function() {
+    this.setState({keyspaceOptions: Object.keys(this.state.clusterTables).map(
+      obj => { return {value: obj, label: obj}; }
+    )});
   },
 
   _onAdd: function(e) {
@@ -123,10 +135,14 @@ const snapshotForm = React.createClass({
       clusterName: this.state.clusterName, 
       owner: this.state.owner,
       name: this.state.name,
-      keyspace: ""
+      keyspace: "",
     };
-    if(this.state.keyspace) snapshot.keyspace = this.state.keyspace;
-    if(this.state.cause) snapshot.cause = this.state.cause;
+    if (this.state.keyspace) {
+      snapshot["keyspace"] = this.state.keyspace;
+    }
+    if (this.state.cause) {
+      snapshot["cause"] = this.state.cause;
+    }
 
     this._takeSnapshot(snapshot);
   },
@@ -139,11 +155,7 @@ const snapshotForm = React.createClass({
     const state = this.state;
     state[n] = v;
     this.replaceState(state);
-    
-    if (n == 'clusterName') {
-      this._getClusterStatus();
-    }
-    
+
     // validate
     this._checkValidity();
   },
@@ -162,58 +174,62 @@ const snapshotForm = React.createClass({
     }
   },
 
-    // Keyspace tag list functions
-    _handleKeyspaceAddition(ks) {
-      let keyspaces = this.state.keyspaceList;
-      if (keyspaces.length==0) {
-        if ($.inArray(ks, this.state.keyspace.split(','))==-1) {
-          keyspaces.push({
-              id: this._create_UUID(),
-              text: ks
-          });
-          this.setState({keyspaceList: keyspaces, keyspace: ks, keyspaces: keyspaces.map(ks => ks.text).join(',')});
-          this._checkValidity();
-          this._getTableSuggestions(ks);
-        }
-      }
-    },
-  
-    _handleKeyspaceDelete(i) {
-        let keyspaces = this.state.keyspaceList;
-        keyspaces.splice(i, 1);
-        this.setState({keyspaceList: keyspaces, keyspace: "", keyspaces: keyspaces.map(ks => ks.text).join(',')});
-        this._checkValidity();
-        this._getTableSuggestions("");
-    },
+  _handleSelectOnChange: function(valueContext, actionContext) {
+    const stateName = actionContext.name.split("_")[1];
+    let stateValue = {};
 
-    _handleKeyspaceFilterSuggestions(textInputValue, possibleSuggestionsArray) {
-      var lowerCaseQuery = textInputValue.toLowerCase();
-      let keyspaces = this.state.keyspaceList;
-   
-      return possibleSuggestionsArray.filter(function(suggestion)  {
-          return suggestion.toLowerCase().includes(lowerCaseQuery) && keyspaces.length==0;
-      })
-    },
+    stateValue[stateName] = valueContext.value;
+    this.setState(stateValue);
 
-    _create_UUID(){
-      var dt = new Date().getTime();
-      var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-          var r = (dt + Math.random()*16)%16 | 0;
-          dt = Math.floor(dt/16);
-          return (c=='x' ? r :(r&0x3|0x8)).toString(16);
+    if (stateName === "clusterName") {
+      this._getClusterStatus();
+    }
+
+    this._checkValidity();
+  },
+
+  _handleKeyspaceSelectOnChange: function(valueContext, actionContext) {
+    let keyspaceListRef = this.state.keyspaceList;
+    let keyspaceRef = "";
+
+    keyspaceListRef.length = 0;
+
+    if (valueContext) {
+      keyspaceListRef.push({
+        id: this._create_UUID(),
+        text: valueContext.value
       });
-      return uuid;
-    },
+      keyspaceRef = valueContext.value;
+    }
+
+    this.setState({
+      keyspace: keyspaceRef,
+      keyspaceList: keyspaceListRef,
+    });
+    this._checkValidity();
+  },
+
+  _create_UUID(){
+    var dt = new Date().getTime();
+    var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+      var r = (dt + Math.random()*16)%16 | 0;
+      dt = Math.floor(dt/16);
+      return (c === 'x' ? r : (r&0x3|0x8)).toString(16);
+    });
+    return uuid;
+  },
 
   render: function() {
+    const clusterNameOptions = this.state.clusterNames.sort().map(name => {
+      return {label: name, value: name};
+    });
 
-    const clusterItems = this.state.clusterNames.sort().map(name =>
-      <option key={name} value={name}>{name}</option>
-    );
+    const clusterNamePlaceholder = this.state.clusterName ? this.state.clusterName : "Select cluster";
 
-    const keyspaceInputStyle = this.state.keyspaceList.length > 0 ? 'form-control-hidden':'form-control';
+    const keyspaceInputStyle = this.state.keyspaceList.length ? 'form-control-hidden':'form-control';
 
-    const form = <div className="row">
+    const form = (
+      <div className="row">
         <div className="col-lg-12">
 
           <form className="form-horizontal form-condensed">
@@ -221,10 +237,14 @@ const snapshotForm = React.createClass({
             <div className="form-group">
               <label htmlFor="in_clusterName" className="col-sm-3 control-label">Cluster*</label>
               <div className="col-sm-9 col-md-7 col-lg-5">
-                <select className="form-control" id="in_clusterName"
-                  onChange={this._handleChange} value={this.state.clusterName}>
-                  {clusterItems}
-                </select>
+                <Select
+                  id="in_clusterName"
+                  name="in_clusterName"
+                  classNamePrefix="select"
+                  options={clusterNameOptions}
+                  placeholder={clusterNamePlaceholder}
+                  onChange={this._handleSelectOnChange}
+                />
               </div>
             </div>
 
@@ -238,20 +258,20 @@ const snapshotForm = React.createClass({
 
             <div className="form-group">
             <label htmlFor="in_keyspace" className="col-sm-3 control-label">Keyspace</label>
-            <div className="col-sm-9 col-md-7 col-lg-5">
-              <ReactTags id={'in_keyspace'} tags={this.state.keyspaceList}
-                suggestions={this.state.keyspaceSuggestions}
-                labelField={'text'} handleAddition={this._handleKeyspaceAddition} 
-                handleInputBlur={this._handleKeyspaceAddition} 
-                handleDelete={this._handleKeyspaceDelete}
-                placeholder={'Empty == all keyspaces'}
-                handleFilterSuggestions={this._handleKeyspaceFilterSuggestions}
-                classNames={{
-                    tagInputField: keyspaceInputStyle
-                  }}/>
+              <div className="col-sm-9 col-md-7 col-lg-5">
+                <Select
+                  id="in_keyspace"
+                  name="in_keyspace"
+                  classNamePrefix="select"
+                  isClearable
+                  isSearchable
+                  options={this.state.keyspaceOptions}
+                  placeholder="select a keyspace otherwise all keyspaces are selected"
+                  onChange={this._handleKeyspaceSelectOnChange}
+                />
               </div>
             </div>
-            
+
             <div className="form-group">
               <label htmlFor="in_owner" className="col-sm-3 control-label">Owner*</label>
               <div className="col-sm-9 col-md-7 col-lg-5">
@@ -259,7 +279,6 @@ const snapshotForm = React.createClass({
                   onChange={this._handleChange} id="in_owner" placeholder="owner name for the snapshot (any string)"/>
               </div>
             </div>
- 
 
             <div className="form-group">
               <label htmlFor="in_cause" className="col-sm-3 control-label">Cause</label>
@@ -274,13 +293,11 @@ const snapshotForm = React.createClass({
                 <button type="button" className="btn btn-warning" disabled={!this.state.submitEnabled}
                   onClick={this._onAdd}>Take snapshot</button>
               </div>
-            </div>            
+            </div>
           </form>
-
+        </div>
       </div>
-    </div>
-
-    
+    );
 
     let menuDownStyle = {
       display: "inline-block" 
@@ -299,21 +316,27 @@ const snapshotForm = React.createClass({
       }
     }
 
-    const formHeader = <div className="panel-title" ><a href="#snapshot-form" data-toggle="collapse" onClick={this._toggleFormDisplay}>Take a snapshot&nbsp; <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuDownStyle}></span><span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuUpStyle}></span></a></div>
+    const formHeader = (
+      <div className="panel-title" >
+        <a href="#snapshot-form" data-toggle="collapse" onClick={this._toggleFormDisplay}>
+          Take a snapshot&nbsp;
+          <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuDownStyle}></span>
+          <span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuUpStyle}></span>
+        </a>
+      </div>
+    );
 
-
-
-
-
-    return (<div className="panel panel-warning">
-              <div className="panel-heading">
-              <NotificationSystem ref="notificationSystem" />
-                {formHeader}
-              </div>
-              <div className="panel-body collapse" id="snapshot-form">
-                {form}
-              </div>
-            </div>);
+    return (
+      <div className="panel panel-warning">
+        <div className="panel-heading">
+          <NotificationSystem ref="notificationSystem" />
+          {formHeader}
+        </div>
+        <div className="panel-body collapse" id="snapshot-form">
+          {form}
+        </div>
+      </div>
+    );
   }
 });
 
