@@ -14,26 +14,39 @@
 //  limitations under the License.
 
 import React from "react";
-import ProgressBar from 'react-bootstrap/lib/ProgressBar';
-import Table from 'react-bootstrap/lib/Table';
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
+import Select from 'react-select';
+import ProgressBar from 'react-bootstrap/ProgressBar';
+import Table from 'react-bootstrap/Table';
 import {DeleteStatusMessageMixin, humanFileSize, getUrlPrefix, toast} from "jsx/mixin";
-import { WithContext as ReactTags } from 'react-tag-input';
 
-const Stream = React.createClass({
+const Stream = CreateReactClass({
     propTypes: {
-        planId: React.PropTypes.string.isRequired,
-        direction: React.PropTypes.string.isRequired,
-        stream: React.PropTypes.object.isRequired
+        planId: PropTypes.string.isRequired,
+        direction: PropTypes.string.isRequired,
+        stream: PropTypes.object.isRequired
     },
 
     getInitialState() {
         return {communicating: false, collapsed: true};
     },
 
-    _tableTags: function(tables) {
-        const tags = tables.reduce((sum, item) => sum.concat({id: item, text: item}), [])
+    _tableList: function(tables) {
+        const selectValues = tables.reduce((sum, item) => sum.concat({value: item, label: item}), [])
         return (
-            <ReactTags tags={tags} readOnly={true} />
+            <Select
+                name="tableList"
+                classNamePrefix="select"
+                value={selectValues}
+                isMulti
+                isClearable={false}
+                isDisabled={true}
+                components={{
+                  DropdownIndicator: () => null,
+                  IndicatorSeparator: () => null
+                }}
+            />
         );
     },
 
@@ -60,7 +73,7 @@ const Stream = React.createClass({
             var directionText = "To: ";
         };
 
-        const tableTags = this._tableTags(tables)
+        const tableList = this._tableList(tables)
 
         const peerWidth = {
             width: "10%"
@@ -79,7 +92,7 @@ const Stream = React.createClass({
             <tr>
                 <td style={peerWidth}> <strong>{directionText} </strong> {stream.peer} </td>
                 <td style={planWidth}> <strong>PlanId: </strong> {this.props.planId} </td>
-                <td style={tableWidth}> <strong>Tables: </strong> {tableTags} </td>
+                <td style={tableWidth}> <strong>Tables: </strong> {tableList} </td>
                 <td style={barWidth}> <ProgressBar now={progress} active={isActive} label={label} bsStyle={style} key={stream.id} /> </td>
             </tr>
         );
