@@ -14,21 +14,23 @@
 //  limitations under the License.
 
 import React from "react";
-import Table from 'react-bootstrap/lib/Table';
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
+import Table from 'react-bootstrap/Table';
 import {DeleteStatusMessageMixin, humanFileSize, getUrlPrefix, toast} from "jsx/mixin";
 import $ from "jquery";
 
-const TpStats = React.createClass({
+const TpStats = CreateReactClass({
     propTypes: {
-      endpoint: React.PropTypes.string.isRequired,
-      clusterName: React.PropTypes.string.isRequired
+      endpoint: PropTypes.string.isRequired,
+      clusterName: PropTypes.string.isRequired
     },
 
     getInitialState() {
       return {tpstats: [], scheduler: {}};
     },
 
-    componentWillMount: function() {
+    UNSAFE_componentWillMount: function() {
       this._collectTpstats();
       this.setState({scheduler : setInterval(this._collectTpstats, 10000)});
     },
